Added manager, renamed setter
The "manager" function will help the user add and remove different recipes, with a dedicated sheet, probably multiple at once.

The setter function was renamed to "starter" since it'll gather all the recipe information inside the current sheet, and start recipeTally.js with the help of a button. It may also handle converting the range objects to 2d arrays so there doesn't need to be too much Range object management.