Added error message to recipeGetter, edited addresses
I added another row on top of the recipe entry & crafting tally sheet, so now the address that they start at are slightly lower inside recipeGetter(), recipeManager() and recipeStarter(). (E1 in the crafting tally sheet, and H1 in the recipe entry sheet)
And whenever recipeGetter() finds a blank range of entries to send to recipeTally(), it sends an error message in the result area instead.