Added sheet name flexibility
Inside recipeManager() and recipeStarter():
I removed most of the static strings to the "stardew" sheet name.
Then I added a reference to a part of the sheet containing a string of a sheet's name to each script where that "stardew" string was. (crafting "E1", recipe "B1", "H1" was a mistake in the last commit)