Moving a few private google sheet scripts
Just a few.
I wanted to share some of my previously-private scripts with people if they need any, and also show that I'm still coding 👍

I'll continue working on quite a few of these since they were made when I was getting used to javascript, and are missing some features.