Remade quickQuery, added simple replacements
I added the start of the new quickQuery function, and it successfully replaces variables named "arr1", "arr2", etc. as the correct variables inside searchFields.

I also added JsDoc documentation at the top of the script, but it's subject to change over time as more functionality is added.

Finally, I added a test function inside a test spreadsheet for future projects so i can rapidly debug the functions later on.