Implemented recipe yield
During the recipe making, it compares the yield of the recipe to the multiplier and takes the ceiling of the needed materials.

For example: in Minecraft, a log outputs 4 planks, so if a recipe uses 6 planks, it should count the log twice (8 planks) when it comes to that point of the recursion instead of something like a fraction/decimal.