Added comments and a minor cost change
The script uses the "bought for" column instead of the "sells for" column for now, and it totals the costs of raw materials instead of how much each would sell for on its own. In the future I'll think more about the "sells for" column since I'll need tom compare most of the ingredients in the tree for the best prices.